THe only  good thing about this pull request is that Controller and View are now interfaces

+++ src/main/models/components/interfaces/Weapon.java
@@ -4,6 +4,6 @@ import main.models.components.Collider;
 import main.utils.Pair;
 
 public interface Weapon {
-    void deploy(Pair<Integer, Integer> startingPoint);
+    void deploy(Pair<Double, Double> startingPoint);
     void checkCollision(Collider entityToCheck);
 }
